val change to float

+++ database/migrations/2024_05_22_213814_create_temps_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('temps', function (Blueprint $table) {
             $table->id();
-            $table->integer('val');
+            $table->float('val');
             $table->string('time');
         });
     }
